==> AllData/2.데이터전처리/1_1.listed_unlisted_concat.ipynb <==
import pandas as pd

listed = pd.read_csv('../1.데이터셋/4.상장_Clean_Feature.csv')
listed

unlisted = pd.read_csv('../1.데이터셋/4.비상장_Clean_Feature.csv')
unlisted

unlisted.columns

all_data = pd.concat([listed, unlisted], axis=0)
all_data

all_data = all_data[
    [
        "회사명",
        "거래소코드",
        "회계년도",
        "CASH FLOW 대 부채비율",
        "CASH FLOW 대 차입금비율",
        "CASH FLOW 대 총자본비율",
        "CASH FLOW 대 매출액비율",
        "유동비율",
        "당좌비율",
        "부채비율",
        "유동부채비율",
        "차입금의존도",
        "순운전자본비율",
        "현금비율",
        "유동자산구성비율",
        "자기자본구성비율",
        "자기자본순이익률",
        "경영자본순이익률",
        "매출액순이익률",
        "총자본사업이익률",
        "총자본영업이익률",
        "매출액영업이익률",
        "금융비용부담률",
        "CROE",
        "총자산증가율",
        "총자본증가율",
        "유동자산증가율",
        "비유동자산증가율",
        "매출액증가율",
        "유형자산증가율",
        "부가가치",
        "총자본투자효율",
        "설비투자효율",
        "부가가치율",
        "이윤분배율",
        "자본분배율",
        "총자산회전율",
        "매출채권회전률",
        "당좌자산회전률",
        "유형자산회전율",
        "경영자본회전률",
        "총자본회전률",
        "비유동자산회전률",
        "유동자산회전률",
        "영업년수",
        "쭈피처",
        "영업투자현금비율",
        "영업재무현금비율",
        "부실판단",
        "도입기",
        "성숙기",
        "성장기",
        "쇠퇴기",
    ]
]

all_data.to_csv('../1.데이터셋/Feature/All_Data.csv', index=False)
