==> AllData/2.데이터전처리/2_1.All_Data_train_test_split.ipynb <==
목표 : 상장기업 train_test_split
=====================

import modules

import pandas as pd
import numpy as np

#워닝 메시지
import warnings
warnings.filterwarnings('ignore')

import os
current_directory = os.getcwd()
print("현재 디렉토리:", current_directory)

from sklearn.model_selection import train_test_split

환경설정

# pd.set_option('display.max_rows', None)
pd.reset_option('display.max_rows')

# pd.set_option('display.max_columns', None)
pd.reset_option('display.max_columns')

pd.set_option('display.max_seq_items', None)

pd.set_option('float_format', '{:.3f}'.format) # 소수점 3개까지 출력

load file

all_data = pd.read_csv("../1.데이터셋/Feature/All_Data.csv")
# 영업년수와 상장년수가 있더라

all_data.columns

# train test split

X = all_data.drop('부실판단', axis=1)
y = all_data['부실판단']
# 데이터를 학습용과 테스트용으로 나눔
all_data_X_train, all_data_X_test, all_data_y_train, all_data_y_test = train_test_split(X, y, test_size= 1 / (np.exp(1) + 1) , random_state=911, stratify=y)

# train data에 부실판단도 넣기
all_data_X_train['부실판단'] = all_data_y_train

# test에도 부실판단 넣기
all_data_X_test['부실판단'] = all_data_y_test
all_data_X_test['부실판단']

# to_csv

# train data 원본
# 결측치나 이상치는 다음에서 처리
all_data_X_train.to_csv('../1.데이터셋/Feature/all_data_train_data.csv', index=False)

all_data_X_test.to_csv("../1.데이터셋/Feature/all_data_test_data.csv", index=False)
