==> AllData/2.데이터전처리/4_5.All_Data_SMOTE oversampling.ipynb <==
목적 : All data Resampling
==============
목표 : 각 수명주기별 SMOTE oversampling
-----------------
각 수명주기별 부실기업 표본수와 정상기업 표본수를 일치하도록 임의로 추출한다.

import modules

import pandas as pd
import numpy as np

#워닝 메시지
import warnings
warnings.filterwarnings('ignore')

import os
current_directory = os.getcwd()
print("현재 디렉토리:", current_directory)

from imblearn.over_sampling import SMOTE

load files

unlisted_winsorized_train = pd.read_csv("../1.데이터셋/Feature/all_data_winsorized_train_data.csv")

# 데이터 확인

unlisted_winsorized_train.columns

unlisted_winsorized_train['부실판단'].value_counts()

unlisted_winsorized_train[['도입기', '부실판단']].value_counts()

unlisted_winsorized_train[['성장기', '부실판단']].value_counts()

unlisted_winsorized_train[['성숙기', '부실판단']].value_counts()

unlisted_winsorized_train[['쇠퇴기', '부실판단']].value_counts()

# Undersampling 표본추출

# SMOTE 방식의 오버샘플링 수행
sampler = SMOTE(sampling_strategy='auto', random_state=42)

# feature와 label을 하나의 csv 파일로 만들예정

도입기 intro

intro_X_train = unlisted_winsorized_train.loc[unlisted_winsorized_train['도입기'] == 1].drop(
    labels=[
        '회사명', '거래소코드', '회계년도'
    ],
    axis=1
)
intro_y_train = unlisted_winsorized_train.loc[unlisted_winsorized_train['도입기'] == 1]['부실판단']
intro_X_resampled_train, intro_y_resampled_train = sampler.fit_resample(intro_X_train, intro_y_train)

intro_X_resampled_train[['도입기', '부실판단']].value_counts()

성장기 growth

growth_X_train = unlisted_winsorized_train.loc[unlisted_winsorized_train['성장기'] == 1].drop(
    labels=[
        '회사명', '거래소코드', '회계년도'
    ],
    axis=1
)
growth_y_train = unlisted_winsorized_train.loc[unlisted_winsorized_train['성장기'] == 1]['부실판단']
growth_X_resampled_train, growth_y_resampled_train = sampler.fit_resample(growth_X_train, growth_y_train)

growth_X_resampled_train[['성장기', '부실판단']].value_counts()

성숙기 maturity

maturity_X_train = unlisted_winsorized_train.loc[unlisted_winsorized_train['성숙기'] == 1].drop(
    labels=[
        '회사명', '거래소코드', '회계년도'
    ],
    axis=1
)
maturity_y_train = unlisted_winsorized_train.loc[unlisted_winsorized_train['성숙기'] == 1]['부실판단']
maturity_X_resampled_train, maturity_y_resampled_train = sampler.fit_resample(maturity_X_train, maturity_y_train)

maturity_X_resampled_train[['성숙기', '부실판단']].value_counts()

쇠퇴기 decline

decline_X_train = unlisted_winsorized_train.loc[unlisted_winsorized_train['쇠퇴기'] == 1].drop(
    labels=[
        '회사명', '거래소코드', '회계년도'
    ],
    axis=1
)
decline_y_train = unlisted_winsorized_train.loc[unlisted_winsorized_train['쇠퇴기'] == 1]['부실판단']
decline_X_resampled_train, decline_y_resampled_train = sampler.fit_resample(decline_X_train, decline_y_train)

decline_X_resampled_train[['쇠퇴기', '부실판단']].value_counts()

# make csv file

concat

train_data = pd.concat([intro_X_resampled_train, growth_X_resampled_train, maturity_X_resampled_train, decline_X_resampled_train], axis=0, ignore_index=True)

train_data[['쇠퇴기', '부실판단']].value_counts()

to_csv

train_data.to_csv("../1.데이터셋/Feature/all_data_smoted_train_data.csv", index=False)
